"NixOS Gem Mismatch" lock gems without installing
`bundle update` updates the lock fine *and* installs
the new gems. This can be problematic if necessary
headers or dependencies are not installed on the
system, which is likely with NixOS where headers are
not generally provided globally.

`bundle lock --update` performs a lock, which
correctly sets the ruby PLATFORMS, without installing.

https://stackoverflow.com/questions/26335777/updating-gemfile-lock-without-installing-gems

This resolved issues locking on my machine.